warning C4267: 'argument': conversion from 'size_t' to 'DWORD', possible loss of data

+++ include/fmt/format-inl.h
@@ -1351,7 +1351,7 @@ FMT_FUNC void vprint(std::FILE* f, string_view format_str, format_args args) {
     internal::utf8_to_utf16 u16(string_view(buffer.data(), buffer.size()));
     auto written = DWORD();
     if (!WriteConsoleW(reinterpret_cast<HANDLE>(_get_osfhandle(fd)),
-                       u16.c_str(), u16.size(), &written, nullptr)) {
+                       u16.c_str(), static_cast<DWORD>(u16.size()), &written, nullptr)) {
       throw format_error("failed to write to console");
     }
     return;
